requirements.txt
cryptography
discord.py
